LayoutTests/inspector/model/remote-object/string.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../../http/tests/inspector/resources/inspector-test.js"></script>
<script src="resources/remote-object-utilities.js"></script>
<script>
function test()
{
    let steps = [
        {expression: `''`},
        {expression: `'"'`}, // a single quoted string
        {expression: `"'"`}, // a double quoted string
        {expression: "`template`"}, // a template string
        {expression: `"string"`},
        {expression: `"Unicode…"`},
        {expression: `"I wish I had something to put here."`},
    ];

    if (!window.WI) {
        window.steps = steps;
        return;
    }

    runSteps(steps);
}
</script>
</head>
<body onload="runTest(); runInBrowserTest();"></body>
</html>
