LayoutTests/inspector/model/remote-object/string-expected.txt


-----------------------------------------------------
EXPRESSION: ''
{
  "_type": "string",
  "_description": "",
  "_value": ""
}

-----------------------------------------------------
EXPRESSION: '"'
{
  "_type": "string",
  "_description": "\"",
  "_value": "\""
}

-----------------------------------------------------
EXPRESSION: "'"
{
  "_type": "string",
  "_description": "'",
  "_value": "'"
}

-----------------------------------------------------
EXPRESSION: `template`
{
  "_type": "string",
  "_description": "template",
  "_value": "template"
}

-----------------------------------------------------
EXPRESSION: "string"
{
  "_type": "string",
  "_description": "string",
  "_value": "string"
}

-----------------------------------------------------
EXPRESSION: "Unicode…"
{
  "_type": "string",
  "_description": "Unicode…",
  "_value": "Unicode…"
}

-----------------------------------------------------
EXPRESSION: "I wish I had something to put here."
{
  "_type": "string",
  "_description": "I wish I had something to put here.",
  "_value": "I wish I had something to put here."
}
